src/backend/pyproject.toml
[project]
name = "fastapi_app"
version = "1.0.0"
description = "Create a RAG application with FastAPI and PostgreSQL"
dependencies = [
    "fastapi~=0.111.0",
    "python-dotenv~=1.0",
    "environs~=11.0",
    "azure-identity~=1.0",
    "aiohttp~=3.0",
    "asyncpg~=0.29.0",
    "SQLAlchemy[asyncio]~=2.0",
    "pgvector~=0.2.0",
    "openai~=1.0",
    "tiktoken~=0.7.0",
    "openai-messages-token-helper~=0.1.0",
    "azure-monitor-opentelemetry~=1.0",
    "opentelemetry-instrumentation-sqlalchemy==0.46b0",
    "opentelemetry-instrumentation-aiohttp-client==0.46b0",
    "opentelemetry-instrumentation-openai~=0.25.0",
    "psycopg2~=2.0",
    "requests~=2.0",
    "sshtunnel~=0.4.0",
    "beautifulsoup4~=4.0",
    "uvicorn~=0.30.0"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
